Add configuration option for craft node hiding
The feature was being complained about and called a bug very often, and because the bug it exists to fix is a minor one, and this itself arguably goes against the "don't edit the vanilla game" aspect of nautilus, it's only right to make it configurable
